+++ server/src/main/java/net/oneandone/stool/server/ui/ApplicationInitializer.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package net.oneandone.stool.server.dashboard;
+package net.oneandone.stool.server.ui;
 
 import org.springframework.security.web.context.AbstractSecurityWebApplicationInitializer;
 import org.springframework.web.context.ContextLoaderListener;