+++ CHANGELOG.md
@@ -8,6 +8,8 @@ it no longer supports shared environments
 New philio: projects are the former stages, this is where your sources reside.
 Stages are now the central part; multiple project can be associated with the same stage.
 
+* dashboard is now part of the server - it's the name over the overview page
+
 * dumped debian packages
 * 1 image per war
 * stool is no longer responsible to manage checkouts - use your standard tools instead;